Dragging, collision, and pausing all successfully implemented.
ERROR: When instantiating the balls and the player, the game gets a null pointer exception coming from the rectangle.intersects method (L:786 of that) in our checkCollision (L:151). Tried moving the balls further from the walls thinking it may have been trying to check collision with rectangles that were outside the frame area, but it still gives the error.